+++ src/vehicle.h
@@ -1070,7 +1070,7 @@ class vehicle
         /*@}*/
 
         // Extra drag on the vehicle from components other than wheels.
-        float drag() const;
+        int drag() const;
 
         // strain of engine(s) if it works higher that safe speed (0-1.0)
         float strain() const;
